% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data Set}
\format{
A data frame with 2 rows and 15 columns:
\describe{
  \item{DISTANCE}{Distance from fire station in miles}
  \item{DAMAGE}{x 1000 dollars in damage}
}
}
\source{
<https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850>
}
\usage{
fire
}
\description{
A data set collected concerning home damage from fire.
}
\keyword{datasets}
