% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iter}

\item{x}{a vector}

\item{fun}{string of what label should be}

\item{alpha}{the alpha}

\item{cx}{the cx}

\item{...}{extra}
}
\value{
a hisogram of Bootstrap sample statistics
}
\description{
myboot2
}
\examples{
myboot2(iter=10000,c(1,1,1,2,2,2,2,3,3,3,4,4),fun="mean",alpha=0.05,cx=1.5)
}
