% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Central limit theorem poisson distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{lambda}{rate parameter}

\item{...}{just an extra}
}
\value{
returns a histogram of sample mean, barplot of sample y, and probability function for poisson
}
\description{
Central limit theorem poisson distribution
}
\examples{
mycltp(2,10000,10)
}
