% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Title}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{sample vector}

\item{mu}{mean}

\item{sig}{sigma}

\item{...}{others}
}
\value{
3D graph
}
\description{
Title
}
\examples{
y = c(10,12,13,15,12,11,10)
mymlnorm(x=y,mu=seq(10,15,length=1000),sig=seq(0.1,3.5,length=1000),lwd=2,labcex=1)
}
