% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{lower tail probability starting point}
}
\value{
prints out lower tail
}
\description{
myncurve
}
\examples{
myncurve(mu=7,sigma=4,a=5)
}
