% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysamples}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Iteration number}

\item{time}{When to release table}
}
\value{
A barplot
}
\description{
mysamples
}
\examples{
mysample(n=1000, iter=30, time=1)
}
