% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myzfunction.R
\name{myzfunction}
\alias{myzfunction}
\title{My z function}
\usage{
myzfunction(x)
}
\arguments{
\item{x}{a quantitative vector of data}
}
\value{
A list containing the z value
}
\description{
My z function
}
\examples{
myzfunction(x = 1:10)
}
