% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats in the flight}

\item{gamma}{Probability a plane is truly overbooked}

\item{p}{Probability of a "show"}
}
\value{
2 plots of Objective function Vs n and list containing nd, nc, N, p and gamma are outputed
}
\description{
Title
}
\examples{
ntickets(N=200, gamma = 0.02, p=0.95)
}
